% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdmxClient.R
\name{getTimeSeriesTable2}
\alias{getTimeSeriesTable2}
\title{get time series and return a data.frame (SDMX v3)}
\usage{
getTimeSeriesTable2(
  provider,
  dataflow,
  key = "",
  filter = "",
  start = "",
  end = "",
  attributes = "all",
  measures = "all",
  updatedAfter = .jnull(),
  includeHistory = FALSE
)
}
\arguments{
\item{provider}{the name of the provider}

\item{dataflow}{dataflow of the time series}

\item{key}{timeseries key}

\item{filter}{optional filter to be applied}

\item{start}{the start time - optional}

\item{end}{the end time - optional}

\item{attributes}{the comma separated list of attributes to be returned - optional, default='all', 'none' for no attributes}

\item{measures}{the comma separated list of measures to be returned - optional, default='all', 'none' for no measures}

\item{updatedAfter}{return only changes after this date - optional}

\item{includeHistory}{include history of revisions - optional, default=false}
}
\description{
Extract a list of time series identified by the parameters provided in input, and return a data.frame as result.
getTimeSeriesTable2(provider, dataflow, key, filter, start, end, attributes, measures, updatedAfter, includeHistory)
}
