% Copyright 2010,2014 Bank Of Italy
%
% Licensed under the EUPL, Version 1.1 or as soon they
% will be approved by the European Commission - subsequent
% versions of the EUPL (the "Licence");
% You may not use this work except in compliance with the
% Licence.
% You may obtain a copy of the Licence at:
%
%
% http://ec.europa.eu/idabc/eupl
%
% Unless required by applicable law or agreed to in
% writing, software distributed under the Licence is
% distributed on an "AS IS" basis,
% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
% express or implied.
% See the Licence for the specific language governing
% permissions and limitations under the Licence.
%
\name{RJSDMX-package}
\title{Gets timeseries from SDMX data Provider}
\description{  This package provides functions to extract timeseries from an SDMX Provider (e.g. ECB,OECD, EUROSTAT) via SDMX Web Service}
\details{\tabular{ll}{
Package: \tab RJSDMX\cr
Type: \tab Package\cr
}
Download Timeseries in local environment using Web Services published by SDMX data Providers.
}
\alias{RJSDMX}
\docType{package}
\keyword{package}
\seealso{\bold{getProviders, getTimeSeries, sdmxHelp}}
\examples{
# not run
# my_ts = getTimeSeries('ECB','EXR.M.USD.EUR.SP00.A')
}
\author{Attilio Mattiocco, Diana Nicoletti, Bank of Italy, IT Support for the Economic Research 
\email{attilio.mattiocco@bancaditalia.it, diana.nicoletti@bancaditalia.it}}
\references{\url{http://sdmx.org/}
\url{http://bancaditalia.it}
\url{http://sdw.ecb.europa.eu/}
\url{http://stats.oecd.org/}}

