% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxTS2DF}
\alias{sdmxTS2DF}
\title{Convert SDMX}
\usage{
sdmxTS2DF(SDMXTS, provider, timevar = "date", numeric = TRUE)
}
\arguments{
\item{SDMXTS}{object returned from \code{getSDMX}}

\item{provider}{?}

\item{timevar}{?}

\item{numeric}{?}
}
\description{
Convert objects returned by \code{getSDMX}
}
\details{
Convert objects returned by \code{getSDMX} into other object types
}
\examples{
df <- sdmxTS2DF()
}
\author{
Bo Werth \email{Bo.WERTH@oecd.org}
}
\seealso{
\code{\link{getSDMX}}, \code{\link{makeSDMXTS}}
}
\keyword{SDMX}
\keyword{zoo}

