\name{setProviderCredentials}
\alias{setProviderCredentials}
\title{set Credentials for authenticating providers}
\usage{
setProviderCredentials(provider, user=.jnull(), pw=.jnull())
}
\arguments{
  \item{provider}{the name of the provider}

  \item{user}{the username}

  \item{pw}{the password}
}
\description{
Sets the credentials for providers that need authentication
}